backend/pyproject.toml
[project]
name = "aigis-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.13.0",
    "bcrypt>=4.1.0",
    "cython>=3.1.3",
    "fastapi>=0.116.1",
    "grandalf>=0.8",
    "ipykernel>=6.30.1",
    "ipython>=9.4.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-deepseek>=0.1.4",
    "langgraph>=0.6.5",
    "opik>=1.8.33",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "ruff>=0.12.10",
    "sqlalchemy>=2.0.43",
    "sqlalchemy-stubs>=0.4",
    "uvicorn>=0.35.0",
    "wikipedia>=1.4.0",
]

[tool.ruff.per-file-ignores]
# Ignore F401 (unused import) errors in any env.py file
"**/env.py" = ["F401"]